fix(anilist): change score type from Float to Int in mutation
AniList API expects Int for scoreRaw parameter, not Float

+++ packages/anilist/src/graphql/mutations.ts
@@ -11,7 +11,7 @@ export const SAVE_MEDIA_LIST_ENTRY = gql`
     $status: MediaListStatus
     $progress: Int
     $progressVolumes: Int
-    $score: Float
+    $score: Int
     $repeat: Int
     $priority: Int
     $private: Boolean
